Adds email alerts
This commit adds the notion of availability to Checks. A Check can now
be available, meaning it's up and reachable or vice versa.

The RequestJob now has the responsibility of updating this value for the
check it's working on.

An AlertMailer is added as well, to notify when availability changes for
a Check. The mailer is triggered by a RequestJob, before it updates the
availability of a Check.

For local development, a standard mailcatcher service is added to
compose and for production emails are delivered using mailgun.