Adds 'status' and 'retries' columns to Check
This is the first step in migrating from a binary view of whether a
check is up or down, to introducing an intermediate state of being in
limbo.

This is to alleviate spurious alert triggers, where a check goes down
and then the minute after goes back up again.

By introducing a number of retries and an intermediate state of 'limbo',
the system should be more robust in only reporting real outages.

This commit also includes a rake task to migrate the value of
'available' to be either a status of up or down.