Updates logic for when alert triggering
This utilises the retries and status column added in 2a507a681b, to
introduce the concept of a check being in 'limbo', which effectively
introduces a grace periode, between when the target of a check is up and
when it goes down, triggering alerts.

Instead of immediately sending alerts on first failure, a check is now
put in a state of limbo, and only when a preset number of retries has
been reached, marked as being down.

A slight modification to the notion of interval is introduced by this
change as well. Before, only stale checks were being queued for
requests, but now all checks either in limbo or strictly down are
included in the batch of requests enqueued every minute.